Iterate over tables instead of indexing
Indexing is theoretically constant time if implemented as reading from a
pointer offset in a contiguous memory block. That satisfies the definition
provided by Serai in https://github.com/serai-dex/serai/issues/328.

Due to this not being true in practice, as pointed out in
https://github.com/kayabaNerve/fcmp-plus-plus/issues/2, this patch mitigates
as instructed.

Fixes https://github.com/kayabaNerve/fcmp-plus-plus/issues/2.