Respond to V-FCMP-VUL-001
This defines a single function to sample points on the embedded elliptic curve.

We do not replace the usage of `to_xy().unwap()` as that is two function calls,
making the only more minimal form a single function call. A single function
would omit the declaration of the panic, requiring a more verbose name, such as
`to_xy_panicking`, removing benefit of a shorter form